refactor: Rename capellamodels to toolmodels
