+++ backend/capellacollab/core/database/models.py
@@ -3,10 +3,10 @@
 
 # These import statements of the models are required and should not be removed! (SQLAlchemy will not load the models otherwise)
 import capellacollab.notices.models
-import capellacollab.projects.capellamodels.backups.models
-import capellacollab.projects.capellamodels.modelsources.git.models
-import capellacollab.projects.capellamodels.modelsources.t4c.models
 import capellacollab.projects.models
+import capellacollab.projects.toolmodels.backups.models
+import capellacollab.projects.toolmodels.modelsources.git.models
+import capellacollab.projects.toolmodels.modelsources.t4c.models
 import capellacollab.projects.users.models
 import capellacollab.sessions.models
 import capellacollab.settings.modelsources.git.models
