+++ backend/capellacollab/projects/toolmodels/modelsources/git/crud.py
@@ -6,9 +6,9 @@ import typing as t
 
 from sqlalchemy.orm import Session
 
-import capellacollab.projects.capellamodels.crud as models_crud
-from capellacollab.projects.capellamodels.models import DatabaseCapellaModel
-from capellacollab.projects.capellamodels.modelsources.git.models import (
+import capellacollab.projects.toolmodels.crud as models_crud
+from capellacollab.projects.toolmodels.models import DatabaseCapellaModel
+from capellacollab.projects.toolmodels.modelsources.git.models import (
     DatabaseGitModel,
     PatchGitModel,
     PostGitModel,
