+++ backend/capellacollab/projects/toolmodels/modelsources/git/routes.py
@@ -9,11 +9,11 @@ from sqlalchemy.orm import Session
 
 from capellacollab.core.authentication.database import ProjectRoleVerification
 from capellacollab.core.database import get_db
-from capellacollab.projects.capellamodels.injectables import (
+from capellacollab.projects.toolmodels.injectables import (
     get_existing_capella_model,
 )
-from capellacollab.projects.capellamodels.models import DatabaseCapellaModel
-from capellacollab.projects.capellamodels.modelsources.git.models import (
+from capellacollab.projects.toolmodels.models import DatabaseCapellaModel
+from capellacollab.projects.toolmodels.modelsources.git.models import (
     DatabaseGitModel,
     PatchGitModel,
     PostGitModel,
