+++ backend/capellacollab/projects/toolmodels/modelsources/t4c/injectables.py
@@ -6,12 +6,12 @@ from sqlalchemy.exc import NoResultFound
 from sqlalchemy.orm import Session
 
 from capellacollab.core.database import get_db
-from capellacollab.projects.capellamodels.injectables import (
+from capellacollab.projects.toolmodels.injectables import (
     get_existing_capella_model,
 )
-from capellacollab.projects.capellamodels.models import DatabaseCapellaModel
-from capellacollab.projects.capellamodels.modelsources.t4c import crud
-from capellacollab.projects.capellamodels.modelsources.t4c.models import (
+from capellacollab.projects.toolmodels.models import DatabaseCapellaModel
+from capellacollab.projects.toolmodels.modelsources.t4c import crud
+from capellacollab.projects.toolmodels.modelsources.t4c.models import (
     DatabaseT4CModel,
 )
